Preserve PATH for Vespa services
common-env.sh sets PATH, instead of adding to it. This means e.g. binaries
reachable via PATH, but not in these whitelisted locations, are lost. For our
tester application we have a custom image containing a custom maven
installation, reachable with the PATH.

The effect of setting PATH seems to have been to add VESPA_HOME/bin[64] only.

To ensure all programs reachable with the earlier code, is still reachable
after this code change, all earlier paths are prepended to PATH.