Improve memory management in all enum attributes.
The new enum store uses 1024 small data buffers instead of 2 large as before.
This avoids the problem with memory spikes when the active buffer was full and
all values had to be compacted into the other buffer.
In addition the new enum store uses free lists such that compaction is not needed as often.