Check target OS version per zone
This change makes the controller check the target OS version per zone (retrieved
from `/nodes/v2/upgrade/`) when deciding whether to set a target.

In the node repository, the OS version value in `/nodes/v2/upgrade/` and the value of
each nodes `wantedOsVersion` are backed by the same ZK data.

To implement automatic pausing of OS upgrades we'll eventually start omitting
the `wantedOsVersion` node field when an OS version is deactivated. We therefore
make the controller read target version from `/nodes/v2/upgrade/` instead to
avoid the controller repeatedly setting a target version while upgrades to that
version are paused.