Inhibit merges when ideal node is unavailable in pending state
Upon entering a cluster state transition edge the distributor will
prune all replicas from its DB that are on nodes that are unavailable
in the _pending_ state. As long as this state is pending, the _current_
state will include these nodes as available. But since replicas for
the unavailable node(s) have been pruned away, started merges that
involve these nodes as part of their chain are doomed to fail. We
therefore inhibit such merges from being started in the first place.