Include retired cluster controllers
When all nodes are replaced we lose the cluster state if we only
allocate cluster controllers on non-retired nodes. This does not
work with global documents as it leads all nodes to enter the
maintenance state where buckets are deactivated until global
documents are completely migrated to new nodes.

By including the retired cluster controllers in the cluster
controller cluster we preserve the existing ZooKeeper state
which prevents this situation.