Fix setup of same element iterator to use the attribute search context from the child blueprint instead of the child search iterator.
This fixes a bug that occurs if the search iterator from a fast-search attribute is a bit vector iterator.
The bit vector iterator doesn't expose the attribute search context, so the setup of the same element iterator
doesn't wrap it into an attribute element iterator that handles finding which elements that match.
The result is that the same element iterator will not match any documents.