Read reboot-interval-in-days dynamically
But also:

Changes the distribution of the scheduling past 1x reboot interval: hosts will
be scheduled for reboot evenly distributed in the whole 1x-2x range, and is by
this guaranteed to be scheduled at latest at 2x.

The expected time before a reboot was scheduled was 1.33 reboot intervals,
while there was no guarantee of an upper time. The new algorithm has an
expected time before reboot of 1.5 reboot intervals, bound to 2x. The old would
have a higher probability of reboot passing the 1x boundary, while a lower
probability than the new as one nears 2x.

So I think the new algorithm also have the nice property of avoiding thundering
herd, perhaps even more so than the old: For instance when most hosts are
rebooted at the same time in a zone, they would tend to be rescheduled for
reboot closer to each other with the old than with the new.

And, enabling the new algoritm should also not lead to too many hosts suddenly
having to reboot, or at least that's what I hope. I can sanity-check this
before merge - I guess it would be dominated by the number of hosts in
west/east that are beyond 2x.