Add configurable support for weakly consistent client Gets
If configured, Get operations initiated by the client are flagged
with weak internal consistency. This allows the backend to bypass
certain internal synchronization mechanisms, which minimizes latency
at the cost of possibly not observing a consistent view of the
document. This config should only be used in a very restricted set
of cases where the document set is effectively read-only, or cross-
field consistency or freshness does not matter.

To enable the weak consistency, use an explicit config override:

```
<config name="vespa.config.content.core.stor-distributormanager">
  <use_weak_internal_read_consistency_for_client_gets>
    true
  </use_weak_internal_read_consistency_for_client_gets>
</config>
```

This closes #11811