Propagate all source and provider properties by default
This changes default behavor from only propagating native properties.
I believe that will probably be safe, and it just seems more correct.

If somebody have problems with it they can recover the old behavior
by using the strict-contracts config. That config is not mentioned
in the documentation and not touched by the model. We should remove
it on Vespa 8.