Remove use-bucket-space-metric feature flag
The flag controlled config read by the Cluster Controller. Therefore, I have
left the ModelContextImpl.Properties method and implementation (now always
returning true), but the model has stopped using that method internally, and
the config is no longer used in the CC.

The field in the fleetcontroller.def is left unchanged and documented as
deprecated.