Correctly handle search for "false" in the underlying search context of singleboolattribute.
This code path is used when searching on an imported bool attribute (from a parent document),
combined with other query terms (e.g. under an AND) such that the search iterator for the imported bool attributes is not strict.
In this case the underlying search context is used, and search for "false" would previously return all "true" documents instead.