Reduce access logging
Avoids writing access logs in various tests.

1. Disables by-default access logging with Application, since it is used in
unit tests.  2. However many tests create additional DeployState which renders
this ineffective, and so this PR also explicitly disables access logging in
services.xml of some tests.

(1) might be unnecessary if we anyway have to do (2) everywhere, but this is
not clear to me.