Resume host on freeze timeout
Observed on host: The host tries to suspend all services on the host, but is
not allowed to do so for a long period. However the host itself is allowed to
be down, and that is not released even during freeze timeouts (when the nodes
are resumed once). This is unfortunate as the host is suspended unnecessarily.