Remove user tenants
@mpolden please review. @tokle FYI.

This removes all traces of user tenants from our controllers.
Existing deployments (there are none right now) are deactivated,
and applications deleted on restart of the controller, for all
applications under a user tenant.
Existing user tenants are deleted on restart of the controller.
APIs pretends creating user tenants is OK, and that any user
tenant already exists, for the purposes of dashboard compatibility.