Define completeness of SuperModel and DuperModel
In order for Orchestrator to remove application data from ZooKeeper, it must
know which applications do NOT exist. Since the duper model starts with 0
applications, always, the only way of knowing what applications do not exist is
for the bootstrap code to notify the super model/duper model when bootstrap is
complete. There are 2 sources of applications that must signal completeness:

 - The super model, once all applications have been redeployed in
   ConfigServerBootstrap.
 - The infrastructure application, in the InfrastructureProvisioner the first
   time it runs.