Throw if accessing duper model while holding status service application lock
When the duper model is updated, ZooKeeper is atomically updated to e.g. remove
extraneous hosts.  This is done by acquiring the duper model lock first, then
the relevant application lock in the status service.

Acquiring these two locks in the reverse order may lead to a deadlock.

This PR throws an IllegalStateException when detecting the current thread is
about to acquire the duper model lock when the current thread has acquired the
application lock.