Safeguard against adding disk bundles to allowed duplicates.
This should not be necessary, because disk bundles should be
removed by getObsoleteFileReferences (as long as they are
present in both the old and new set of bundles). However, there
was an incident where all the model amender's bundles were added
to the set of allowed duplicates. One possible explanation is that
FileReference.equals or hashCode was broken, so an improved
version is added by this commit.