Support forgetting nodes completely
Normally we want to remember removed nodes by putting them in the
deprovisio´ned state. But if we are
sure they will never re-enter the system we should forget them
completetely. With this that is achieved by issuing another
DELETE on the node when it is in the deprovisioned state.