Add tracking of remove operations rate and use this to consider blocking lid space compaction.
During a period with a high rate of remove operations, there is no use running lid space compaction
as this will interfere with the remove operations, increasing latency of those.
Moving a document as part of lid space compaction is a costly operation (similar to putting the document in the first place)
and it typically uses both the index and attribute writer thread pools.