Mount /etc/machine-id and /var/log/journal
When /var/log/journal is mounted explicitly (/var/log is already mounted, but
that's not sufficient), Docker/systemd will not mount the container's
/var/log/journal/MACHINE-ID on the host's /var/log/journal/MACHINE-ID.

Also, upgrades and restarts of a container will now reuse the same machine ID,
as /etc/machine-id is written once and mounted in.  Dirtying the container node
means a new machine ID must be written for the same hostname later.