Delete expired non-active sessions
Due to a bug fixed in https://github.com/vespa-engine/vespa/pull/13338
we ended up with sessions that are in state ACTIVATED in zookeeper,
but that was not the latest active session.  These would not be
cleaned up, the changes here will delete these sessions (if they are
expired and not the current active session for an application).

Also deletes a test that was covered by tests in ApplicationRepositoryTest.