Run prepare step of two-phase puts in the shared executor instead of the attribute field executor.
Benchmarking has shown that we get an uneven balance between the underlying threads when using the attribute field executor.
This is a sequenced task executor, where each underlying thread only is allowed to handle a subset of the prepare tasks (based on the local document id).
Using the shared executor should give an even balance.