Preserve allocation when possible
We want to avoid unnecessary small adjustments to capacity.
Since the worst case real resources we'll get is generally worse
than the actual allocation, we cannot guarantee that we compute
the same allocation to request as the one leading to the current
allocation. So instead of relying on that we explicitly reuse
the current (requested) resources when they are still legal
instead of recomputing each time.