Ensure hold lists are empty upon DB destruction
If there was a snapshot reader concurrent with the last modify operation
on the DB, it's possible for the hold list to be non-empty. Explicitly
clean it up now to ensure that we don't try to destroy any data stores
with a non-empty hold list. Failure to do so might trigger an assertion.