Deadlock detection
Just before Lock.acquire() is invoked, the locks within the process is queried
to see if a "deadlock" will occur: The current thread waiting to acquire lock
path P1, which is held by thread T1 waiting on acquiring a lock at path P2,
etc, until a thread is waiting for a lock held by the current thread.

Even without this PR the deadlock would resolve itself automatically because
all locks are acquired with timeouts.  However, this PR 1. resolves the
deadlock immediately, and 2. leaves a log trace (hopefully from the exception)
to allow us to refactor code to avoid such deadlocks.