Defer changes until restart
In container clusters, if it is determined that a restart is required to swicth
to a new config generation, mark that config as applicable only after a restart.

This avoids wasting resources on a config change just before a restart, and the
potential of prematurely applying config changes which depend on other changes
which are not effective before a restart has been done.