Add striped implementation of B-tree content node bucket database
Abstracts away multiple underlying B-tree DBs that each hold a subset
of the super bucket space. Offers ordered iteration via a priority-queue
based view over the sub DBs.

Not yet ready for prime time, as the striping inherently requires an
absolute lower bound on the bucket bits used in the system, which is
currently not enforced.