Introduce node IP address pool
A host that is supposed to run containers has a non-empty IP pool: A set of
IPv4 and/or IPv6 addresses that can be assigned to containers.

This PR adds a list of hostnames to this pool.  The intent is that the
hostnames and IPs match through resolving, but resolution may not yet be
available (until DNS changes propagate).

For now, only a list of hostnames are specified.  We may want to specify
(hostname, IP address) pairs or (hostname, IPv4, IPv6) triplets later, and the
serialization format allows for that by storing the hsotnames in an array of
objects, the object having a "hostname" field.

However the REST API is kept simpler for now: it exposes and allows patching of
an array of strings of a "additionalHostnames" field.