Allow allocating to a provisioned tenant host
This PR changes (A) and (B) described below, which effectivelly allows
the allocation of nodes to a dynamically provisioned tenant host.

A.  In NodePriotizer, when converting a VirtualNodeCandidate to a
ConcreteNodeCandidate with 'withNode()', primary IP addresses are currently
picked from the IP pool of the parent host and assigned to the node, along with
the corresponding hostname found using a DNS resolver.

This PR allows a ConcreteNodeCandidate to be created WITH a hostname, but
WITHOUT primary IP addresses, when 1. the parent host has no IP addresses in
the pool AND 2. there are hostnames in the address pool not yet assigned to any
child node.

This may happen for a brief period of time just after provisioning in a
dynamically provisioned zone: An asynchronous process is supposed to add IP
addresses to the pool (based on the pool hostnames) before the host can become
active, and also update the child nodes with their primary IP addresses
accordingly.  They both hold the unallocated lock to guarantee atomicity.

B.  In NodePriotizer.addCandidatesOnExistingHosts(), the
HostCapacity.hasCapacity() will now return true when (a) the parent host is
tenant, (b) there are NO IP addresses in the pool, and (c) there is a hostname
in the pool not assigned to