Make restart on deploy sticky
When restartOnDeploy is set on a config generation, no further config generations
should be applied before the node has been restarted. This is necessary because
the decision on whether restartOnDeploy shouild be set is made by comparing the
current and next config generation on the config server, not the next and the
current one active on the node.