Initial support for backend reindexing visitor functionality
Introduces the concept of a read-for-write visitor operation which
blocks all mutating operations from starting for a bucket being
visited. This read-for-write mode is used if (and only if) the visitor
library being specified by the client is "reindexingvisitor".

Since read-for-write visitors cannot race with concurrent write
operations, starting such visitors are deferred until no further
mutations are pending.

Also adds a basic reindexingvisitor implementation to the content node
which sends all documents as Puts containing a special TaS token
that will let the operation through even if a bucket is locked. This
token is cleared by the distributor before it is passed on to the
content nodes.

Note: this feature is not yet production ready. For now the following
caveats apply:
* Mutating vs non-mutating pending ops to a bucket are not tracked
  separately, so it’s possible to starve a reindexing visitor by
  sending constant pending read load, as read load is not blocked by
  the operation sequencer.
* Ideal state operations towards locked buckets are not blocked, so
  it's possible for e.g. a split to be sent for a bucket that is being
  visited.