Report 0 metrics once
A metric like configserver.lockAttempt.acquireTimedOut.average may be reported
as non-zero once, and then not reported for a while.  In the meantime, the
value reported for this metrics becomes 'last', carried over from earlier
periods.

This PR ensures that the value is set to 0 the first period after the value
becomes 0, but not thereafter (as carrying 0 is fine).