Make it possible to specify feed block resource limits in content cluster tuning.
These will be used by the cluster controller to decide whether
external feed should be blocked and stopped by the distributors.
Content node (proton) hard limits are derived from the new limits if not specified, and the other way around.
This way we are backwards compatible when the new functionality in cluster controller is enabled by default.