Allow only one node down at a time for cluster controller clusters
If we are replacing nodes and have 3 old and 3 new nodes in a cluster,
allowing 50 per cent down would lead to 3 being allowed to go down.
When deploying to remove the 3 old nodes there might be (for a short time)
config that says that there should be 6 nodes in the zookeeper cluster.
This mean 4 of them need to be up for the cluster to have a quorum.
This will not work in this case.